refactor(assess): split assess into stages
