+++ src/assess/README.md
@@ -1,0 +1,12 @@
+# Assess
+
+This is the core of Browsability.  
+
+```txt
+   extract    map   compare
+Files -> Code -> JSON -> Support
+```
+
+1. Extract: retrieve just the relevant code for each analyser from each file
+2. Map: transform the code into a key-value object to make support lookup easier
+3. Compare: find the relevant compatibility data for each key-value pair and aggregate the maximum
